refactor(comment): 패키지 단순화
