--- src/main/java/com/sejong/metaservice/application/comment/command/UpdateCommentCommand.java
@@ -1,5 +1,0 @@
-package com.sejong.metaservice.application.comment.command;
-
-public class UpdateCommentCommand {
-
-}
