+++ src/main/java/com/sejong/metaservice/domain/comment/command/UpdateCommentCommand.java
@@ -1,0 +1,5 @@
+package com.sejong.metaservice.domain.comment.command;
+
+public class UpdateCommentCommand {
+
+}
