+++ src/test/java/com/sejong/metaservice/application/config/MockBeansConfig.java
@@ -2,7 +2,7 @@ package com.sejong.metaservice.application.config;
 
 import static org.mockito.Mockito.mock;
 
-import com.sejong.metaservice.application.comment.service.CommentService;
+import com.sejong.metaservice.domain.comment.service.CommentService;
 import com.sejong.metaservice.domain.like.service.LikeService;
 import org.springframework.context.annotation.Bean;
